Rename file name of SEO component
