+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-import SEO from "../components/SEO/seo"
+import SEO from "../components/SEO/SEO"
 import Landing from "../components/Landing/Landing"
 
 const IndexPage = () => (
